requirements.txt
pandas
requests
pdfplumber
chromadb
google-genai
python-dotenv
